python/requirements.txt
numpy
matplotlib
pandas
Pillow